refactor: refine component code
