+++ apps/web/src/app/event/[category]/components/EventItemSection.tsx
@@ -1,9 +1,10 @@
+'use client';
 import Chip from '@/components/Chip';
 import EventItemCard from '@/components/EventItemCard';
 import { pyeonImage } from '@/dummy/image';
 import React from 'react';
 
-const eventCategory = ['1+1', '2+1', '할인', '덤'];
+const eventCategory = ['1+1', '2+1', '할인', '+덤'];
 
 const EventItemSection = () => {
   return (
@@ -14,7 +15,12 @@ const EventItemSection = () => {
       <div className="pb-[29px]">
         <Chip>
           {eventCategory.map((item, index) => (
-            <Chip.Item myIndex={index} key={index} text={item} />
+            <Chip.Item
+              myIndex={index}
+              key={index}
+              text={item}
+              onClickChipItem={() => console.log('클릭!')}
+            />
           ))}
         </Chip>
       </div>
