+++ apps/web/src/app/event/[category]/components/SearchIconButton.tsx
@@ -1,0 +1,12 @@
+import SearchIcon from '@/components/icons/SearchIcon';
+import React from 'react';
+
+const SearchIconButton = () => {
+  return (
+    <button className="mr-[_-8px] p-[8px]">
+      <SearchIcon type="black" />
+    </button>
+  );
+};
+
+export default SearchIconButton;
