+++ apps/web/src/components/icons/HomeIcon.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Home from '../../assets/home.svg';
+import Home from 'public/icon/home.svg';
 
 const HomeIcon = () => {
   return <Home />;
